Split mysql storage test into 3 files, a base node test + log and map
I think there's minimal overlap now. There's no functionality changes,
just moving code around.
